WIP: AgentAction improvements, dataStream.
Currently I am not sure whether to put dataStream on EndOfData or rename EndOfData to DataStream or what.

I feel confident that it's best to spool all the data from the client into a reliable stream for the server to process (e.g. verify signatures, etc). However, it's unclear whether I should attach that to the command.